Add jmx.rmiregistry.port property to documentation
The `rmiregistry` port is a required property to be able to connect to the JMX server. The documentation currently only mentions the `rmiserver` port, which the registry uses to redirect client requests, but clients actually connect to the `rmiregistry` port.

I have completed the Facebook CLA for this GitHub account, but this is my first contribution.